m4boot: add vf610m4bootldr to get FDT pointer from GPR3
The boot loader running on Cortex-A5 passes the pointer to the device
tree file using the SoC register SRC_GRP3 (argument register for
secondary core). In order to get this pointer to r2 of the Cortex-M4
we need a very small boot loader running just before the kernel starts
booting.